Finished compile pass tests for literal visitors. Still need to finish pass tests for enum visitors, and fail tests for both literal and enum visitors.
