+++ tests/compile-fail/literal-null.missing-handler.ts
