+++ tests/compile-fail/literal-null.missing-null-handler.ts
