+++ tests/compile-fail/literal.missing-handler.ts
