+++ tests/compile-fail/literal.with-null-handler.ts
