+++ tests/compile-pass.literal-both.test.ts
@@ -1,0 +1,3 @@
+import { testCompilation } from "./util/testCompilation";
+
+testCompilation(testCompilation.VisitorType.LiteralWithBoth, true);
