+++ tests/compile-pass.literal-null.test.ts
@@ -1,0 +1,3 @@
+import { testCompilation } from "./util/testCompilation";
+
+testCompilation(testCompilation.VisitorType.LiteralWithNull, true);
