+++ tests/compile-pass.literal-undefined.test.ts
@@ -1,0 +1,3 @@
+import { testCompilation } from "./util/testCompilation";
+
+testCompilation(testCompilation.VisitorType.LiteralWithUndefined, true);
