+++ tests/compile-pass/literal-both.explicit-return-type.ts
@@ -1,0 +1,24 @@
+import visitString from "../../src";
+
+type RGB = "r" | "g" | "b";
+
+const result = visitString("r" as RGB | null | undefined).with<number>({
+    "r": () => {
+        return 1;
+    },
+    "g": () => {
+        return 2;
+    },
+    "b": () => {
+        return 3;
+    },
+    handleNull: () => {
+        return 0;
+    },
+    handleUndefined: () => {
+        return 0;
+    }
+});
+
+// tslint:disable-next-line:no-unused-variable
+const test: number = result;
