+++ tests/compile-pass/literal-null.explicit-param-types.ts
