+++ tests/compile-pass/literal-null.explicit-return-type.ts
