+++ tests/compile-pass/literal-null.inferred-param-type.ts
