+++ tests/compile-pass/literal-null.inferred-return-type.ts
