+++ tests/compile-pass/literal-undefined.inferred-param-type.ts
@@ -1,0 +1,22 @@
+import visitString from "../../src";
+
+type RGB = "r" | "g" | "b";
+
+visitString("r" as RGB | undefined).with({
+    "r": (value) => {
+        // tslint:disable-next-line:no-unused-variable
+        const test: "r" = value;
+    },
+    "g": (value) => {
+        // tslint:disable-next-line:no-unused-variable
+        const test: "g" = value;
+    },
+    "b": (value) => {
+        // tslint:disable-next-line:no-unused-variable
+        const test: "b" = value;
+    },
+    handleUndefined: (value) => {
+        // tslint:disable-next-line:no-unused-variable
+        const test: undefined = value;
+    }
+});
