+++ tests/compile-pass/literal.explicit-param-type.ts
