+++ tests/compile-pass/literal.explicit-return-type.ts
