+++ tests/compile-pass/literal.inferred-param-type.ts
