+++ tests/compile-pass/literal.inferred-return-type.ts
