+++ tests/ts-string-visitor.enum.test.ts
