+++ tsconfig.json
@@ -21,6 +21,6 @@
   },
   "exclude": [
     "node_modules",
-    "__tests__/compile_fail_samples"
+    "tests/compile-fail"
   ]
 }
